update: 인증 필요 여부에 따라 댓글 service 분리
